Decouple linter module from cache module
